requirements.txt
Flask
iso8601
Flask-SQLAlchemy
Flask-Cors
gunicorn
sqlalchemy_utils
python-decouple
psycopg2-binary
flask-restless
passlib
flask-principal
flask-login
flask-marshmallow
marshmallow-sqlalchemy
mock
SQLAlchemy-serializer
pytest
